--- a/jobs/jobs.go
+++ b/jobs/jobs.go
@@ -31,14 +31,14 @@ const (
 type Job struct {
 	ID                     uuid.UUID      `gorm:"column:id;primary_key;type:uuid;"`
 	Type                   string         `gorm:"column:type"`
-	State                  State          `gorm:"column:state;default:INIT"`
+	State                  State          `gorm:"column:state;default:INIT;index:idx_jobs_state_updated_at"`
 	Error                  string         `gorm:"column:error"`
 	Errors                 pq.StringArray `gorm:"column:errors;type:text[]"`
 	Result                 string         `gorm:"column:result"`
 	TransactionID          string         `gorm:"column:transaction_id"`
 	ExecCount              int            `gorm:"column:exec_count;default:0"`
 	CreatedAt              time.Time      `gorm:"column:created_at"`
-	UpdatedAt              time.Time      `gorm:"column:updated_at"`
+	UpdatedAt              time.Time      `gorm:"column:updated_at;index:idx_jobs_state_updated_at"`
 	DeletedAt              gorm.DeletedAt `gorm:"column:deleted_at;index"`
 	ShouldSendNotification bool           `gorm:"-"` // Whether or not to notify admin (via webhook for example)
 	Attributes             datatypes.JSON `gorm:"attributes"`