related: https://github.com/flow-hydraulics/flow-wallet-api/issues/272

Added composite index for jobs table `idx_jobs_state_updated_at(state, updated_at)` to deal with slow query.

### Result
I have confirmed that adding `idx_jobs_state_updated_at(state, updated_at)` to jobs table increases query performance in the following query.

```sql
SELECT
*
FROM
  "jobs"
WHERE
  ((state IN ('INIT', 'ACCEPTED') AND updated_at < '2022/02/05 10:00:00')
    OR (state IN ('ERROR', 'NO_AVAILABLE_WORKERS')
    AND updated_at < '2022/02/04 10:00:00'))
  AND "jobs"."deleted_at" IS NULL
ORDER BY
  created_at DESC
LIMIT 10;
```

#### before
```sql
QUERY PLAN
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=1783.87..1783.88 rows=1 width=131) (actual time=13.793..13.824 rows=0 loops=1)
   ->  Sort  (cost=1783.87..1783.88 rows=1 width=131) (actual time=13.782..13.800 rows=0 loops=1)
         Sort Key: created_at DESC
         Sort Method: quicksort  Memory: 25kB
         ->  Seq Scan on jobs  (cost=0.00..1783.86 rows=1 width=131) (actual time=13.715..13.721 rows=0 loops=1)
               Filter: ((deleted_at IS NULL) AND (((state = ANY ('{INIT,ACCEPTED}'::text[])) AND (updated_at < '2022-02-05 10:00:00+00'::timestamp with time zone)) OR ((state = ANY ('{ERROR,NO_AVAILABLE_WORKERS}'::text[])) AND (updated_at < '2022-02-15 10:00:00+00'::timestamp with time zone))))
               Rows Removed by Filter: 55852
 Planning Time: 55.115 ms
 Execution Time: 13.999 ms
(9 rows)
```

#### after
```sql
QUERY PLAN          
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=21.23..21.24 rows=1 width=135) (actual time=0.154..0.215 rows=0 loops=1)
   ->  Sort  (cost=21.23..21.24 rows=1 width=135) (actual time=0.141..0.190 rows=0 loops=1)
         Sort Key: created_at DESC
         Sort Method: quicksort  Memory: 25kB
         ->  Bitmap Heap Scan on jobs  (cost=17.20..21.22 rows=1 width=135) (actual time=0.058..0.096 rows=0 loops=1)
               Recheck Cond: (((state = ANY ('{INIT,ACCEPTED}'::text[])) AND (updated_at < '2022-02-05 10:00:00+00'::timestamp with time zone)) OR ((state = ANY ('{ERROR,NO_AVAILABLE_WORKERS}'::text[])) AND (updated_at < '2022-02-15 10:00:00+00'::timestamp with time zone)))
               Filter: (deleted_at IS NULL)
               ->  BitmapOr  (cost=17.20..17.20 rows=1 width=0) (actual time=0.044..0.072 rows=0 loops=1)
                     ->  Bitmap Index Scan on idx_jobs_state_updated_at  (cost=0.00..8.60 rows=1 width=0) (actual time=0.016..0.022 rows=0 loops=1)
                           Index Cond: ((state = ANY ('{INIT,ACCEPTED}'::text[])) AND (updated_at < '2022-02-05 10:00:00+00'::timestamp with time zone))
                     ->  Bitmap Index Scan on idx_jobs_state_updated_at  (cost=0.00..8.60 rows=1 width=0) (actual time=0.010..0.015 rows=0 loops=1)
                           Index Cond: ((state = ANY ('{ERROR,NO_AVAILABLE_WORKERS}'::text[])) AND (updated_at < '2022-02-15 10:00:00+00'::timestamp with time zone))
 Planning Time: 3.143 ms
 Execution Time: 0.423 ms
(14 rows)
```