@latenssi 2022-02-16
Hi @seita-uc! Awesome work, thank you! That is quite an impressive improvement. 

Only one note: the Job struct in the migrations file should be a copy of the current Job struct in the application code. So instead of changing only the migrations Job, you should change the struct in here: [jobs/jobs.go](https://github.com/flow-hydraulics/flow-wallet-api/blob/main/jobs/jobs.go) and copy it over to the migration file.